chore: fix and unskip tasks tests
